Moved themes into separate files and switched to dark theme
