+++ web/src/helpers/themes.js
@@ -1,0 +1,21 @@
+import { dark as defaultDark, grommet as defaultLight } from 'grommet/themes'
+
+const theme = {
+    global: {
+      colors: {
+        brand: '#FDA7DF',
+      },
+      font: {
+        family: 'Roboto',
+        size: '14px',
+        height: '20px'
+      }
+    }
+  };
+
+const dark = defaultDark;
+const light = defaultLight;
+
+export {
+    dark, light
+}
